+++ activejob/test/jobs/association_loading_job.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+class AssociationLoadingJob < ArgumentsRoundTripJob
+  def perform(record, loaded_associations = [])
+    loaded_associations.each do |loaded_association|
+      Array(record.public_send(loaded_association))
+    end
+
+    super
+  end
+end